Reorganize protocol code
